큐로 슬라이딩 윈도우 (어려웠음 복습필요)

+++ 2. 자료구조/2-1. 스택, 큐/JuHan_boj_3078.java
@@ -11,7 +11,7 @@ import java.util.StringTokenizer;
 public class Main {
     static int N, K;
     static Queue<Integer> Q[];
-    static int ans;
+    static long ans;
 
     public static void main(String[] args) throws IOException {
         System.setIn(new FileInputStream("src/WEEK15/P3078/input.txt"));
@@ -31,14 +31,14 @@ public class Main {
 
             if(Q[len].isEmpty()){     // 큐가 비었다면 걍 넣음
                 Q[len].add(i);
-            } else {                // 큐를 투포인터처럼 사용
+            } else {                // 아니라면 가장 앞의 등수를 확인
                 while(!Q[len].isEmpty()){
-                    if(i - Q[len].peek() <= K){
+                    if(i - Q[len].peek() <= K){     // 현재 학생의 등수의 차가 K 이하라면
                         ans += Q[len].size();
                         Q[len].add(i);
                         break;
                     }
-                    // 만약 큐의 가장 앞에 있는 등수와의 차가 K보다 크다면 빼낸다.
+                    // 만약 큐의 가장 앞에 있는 등수와의 차가 K보다 크다면 K 이하가 나올때까지 빼낸다.
                     Q[len].poll();
                 }
                 // 예외 : 큐가 비었다면 큐에 있었던 모든 경우의 수의 차가 K보다 큰 것이다.
